+++ packages/core/src/index.ts
@@ -2,6 +2,7 @@ export * from "./address.ts";
 export * from "./abort-signal.ts";
 export * from "./attachment.ts";
 export * from "./message.ts";
+export * from "./message-id.ts";
 export * from "./priority.ts";
 export * from "./receipt.ts";
 export * from "./transport.ts";